+++ lib/td_dd_web/views/data_structure_link_view.ex
@@ -2,6 +2,7 @@ defmodule TdDdWeb.DataStructureLinkView do
   use TdDdWeb, :view
 
   alias TdDd.DataStructures.DataStructure
+  alias TdDdWeb.LabelView
 
   def render(
         "bulk_create.json",
@@ -33,6 +34,13 @@ defmodule TdDdWeb.DataStructureLinkView do
 
   def render("show.json", %{data_structure_link: link}) do
     %{data: render_one(link, __MODULE__, "data_structure_link.json")}
+  end
+
+  def render("search.json", %{result: %{data_structure_links: links, scroll_id: scroll_id}}) do
+    %{
+      data_structure_links: render_many(links, __MODULE__, "data_structure_link_search.json"),
+      scroll_id: scroll_id
+    }
   end
 
   def render("data_structure_link.json", %{
@@ -74,5 +82,18 @@ defmodule TdDdWeb.DataStructureLinkView do
         %{data_structure_link: link}
       ) do
     Map.take(link, [:source_id, :target_id, :inserted_at])
+  end
+
+  def render(
+        "data_structure_link_search.json",
+        %{data_structure_link: link}
+      ) do
+    labels =
+      LabelView.render("index.json", %{labels: link.labels})
+      |> Map.get(:data)
+      |> Enum.map(&Map.get(&1, :data))
+
+    Map.take(link, [:id, :source_id, :target_id, :inserted_at, :updated_at])
+    |> Map.put(:labels, labels)
   end
 end