+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - [TD-7192] Use whitespace tokenizer with word delimiter filter to ignore `_` on text search
 
+## [Unreleased]
+
+### Added
+
+- [TD-7115] Add Data Structure Link incremental search
+
 ## [7.3.1] 2025-03-26
 
 ### Added